fix(maintenance): add bootstrap.tsx and import from index
